requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
